Add some minor fixes to listTable
Ahora no se puede eliminar un plato que ha sido confirmado por el cocinero.
Para evitar problemas con la base de datos.

+++ src/routes/api/deletePlatoPedidoJS.js
@@ -8,32 +8,51 @@ export async function post({ request }) {
 
     try {
 
-        const result = await prisma.plato_Pedido.deleteMany({
+        let result2 = await prisma.plato_Pedido.findFirst({
             where: {
                 plato_id: json.id
             }
         })
 
-        let body;
-
-        if (!result) {
+        if(result2.estado=="En_Proceso"){
+            const result = await prisma.plato_Pedido.deleteMany({
+                where: {
+                    plato_id: json.id
+                }
+            })
+    
+            let body;
+    
+            if (!result) {
+                body = {
+                    message: "Error deleting pedido",
+                    status: 500,
+                };
+                // return validation errors
+                return {
+                    body,
+                };
+            } else {
+                body = {
+                    message: "pedido deleted",
+                    status: 200,
+                };
+                return {
+                    body,
+                };
+            }
+        }else{
             body = {
-                message: "Error deleting pedido",
-                status: 500,
-            };
-            // return validation errors
-            return {
-                body,
-            };
-        } else {
-            body = {
-                message: "pedido deleted",
-                status: 200,
+                message: "pedido no esta en proceso, está confirmado",
+                status: 403,
             };
             return {
                 body,
             };
         }
+
+
+
     } catch (errors) {
         console.log(errors);
         body = {
